+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "crates" extension will be documented in this file.
 
 ### 0.2.1
 * Patreon link changed to a badge.
+* Marketplace badges are linked to marketplace. [#13](https://github.com/serayuzgur/crates/issues/13).
 
 
 ### 0.2.0