+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 All notable changes to the "crates" extension will be documented in this file.
 
+## 0.2.2
+* "-rc" added for pre-release control.
 
 ### 0.2.1
 * Patreon link changed to a badge.